migrate calculate distance functions to a new file
