+++ benchmarks/transport_montecarlo_packet.py
@@ -5,7 +5,6 @@ Basic TARDIS Benchmark.
 import numpy as np
 from asv_runner.benchmarks.mark import parameterize, skip_benchmark
 
-import tardis.transport.frame_transformations as frame_transformations
 import tardis.transport.geometry.calculate_distances as calculate_distances
 import tardis.transport.montecarlo.numba_interface as numba_interface
 import tardis.transport.montecarlo.opacities as opacities
@@ -55,56 +54,6 @@ class BenchmarkMontecarloMontecarloNumbaPacket(BenchmarkBase):
             bf_heating_estimator=np.empty((0, 0), dtype=np.float64),
             stim_recomb_cooling_estimator=np.empty((0, 0), dtype=np.float64),
             photo_ion_estimator_statistics=np.empty((0, 0), dtype=np.int64),
-        )
-
-    @parameterize(
-        {
-            "Packet params": [
-                {"mu": 0.3, "r": 7.5e14},
-                {"mu": -0.3, "r": 7.5e13},
-                {"mu": -0.3, "r": 7.5e14},
-            ]
-        }
-    )
-    def time_calculate_distance_boundary(self, packet_params):
-        mu = packet_params["mu"]
-        r = packet_params["r"]
-
-        calculate_distances.calculate_distance_boundary(
-            r, mu, self.geometry.r_inner[0], self.geometry.r_outer[0]
-        )
-
-    @parameterize(
-        {
-            "Parameters": [
-                {
-                    "packet": {"nu_line": 0.1, "is_last_line": True},
-                },
-                {
-                    "packet": {"nu_line": 0.2, "is_last_line": False},
-                },
-            ]
-        }
-    )
-    def time_calculate_distance_line(self, parameters):
-        packet_params = parameters["packet"]
-        nu_line = packet_params["nu_line"]
-        is_last_line = packet_params["is_last_line"]
-
-        time_explosion = self.model.time_explosion
-
-        doppler_factor = frame_transformations.get_doppler_factor(
-            self.static_packet.r, self.static_packet.mu, time_explosion, False
-        )
-        comov_nu = self.static_packet.nu * doppler_factor
-
-        calculate_distances.calculate_distance_line(
-            self.static_packet,
-            comov_nu,
-            is_last_line,
-            nu_line,
-            time_explosion,
-            False
         )
 
     @parameterize(
